Make less calls to the backend
This commit ensures we make less calls to the backend for
experience-tree-node-score instances by detecting we requested for an
entity which does not exist before.  Whenever such an entity does not
exist, we can safely assume we don't need to request it again in the
future.  It will, for as much as we care, not appear magically in the
future.