Make widget.js take source domain into account
It is hard to figure out where the widget.js script is loaded from,
but it's much much easier to figure out where the widgets themselves
should be hosted from.  Given that we can currently only render one
kind of widget sources and that the type of the widget is supplied in
the script, we can scan for the scripts and read the src content.
This in turn allows us to determine the base domain and render the
correct widget.  Win!