HEAD is implicitly created on modern browsers no custom code needed
We expected to need a custom HEAD in cases where the widget would be
embedded and no HEAD tag would exist.  This is a really unexpected
situation, but it might happen.  Modern browsers create the HEAD tag
imiplicitly.  Therefore we don't need to ship code to take this case
into account.