Basic inheritance for widget
The widget now has basic inheritance.  A scorable has been
introduced.  This means that there's currently no more direct
limitation in the experiences at the PointOfInterest.  A widget is
rendered for the PointOfInterest rather than for an individual
experience.

As an extra, we've removed the duplication of the
embeddable-widget and the widgets/show route.

As a minus, we've temporarily removed the caching system for scoring
which should be added again in the future.