Precautionary adapter for point-of-interest.images
The data entry application has this adapter to ensure at least 100
images can be fetched.  Adding this here so we're sure this will happen
also if we wouldn't fetch this up front.

It's debatable whether we want this in here or not.  It should be unused
code.  It is more of a safty measure.