Render the title of a node when only a comment exists
We had decided to render the title of the node when only a comment
existed earlier on.  It didn't matter too much, but the title wasn't
rendered now.  Rendering only the comment was hard in terms of design
and with this approach we skip that and get a cleaner view.  That counts
as yay.